Start extracting
