+++ src/main/python/plotlyst/view/widget/scene/structure.py
@@ -23,20 +23,19 @@ from functools import partial
 from typing import Optional, List, Dict
 
 import qtanim
-from PyQt6.QtCore import Qt, pyqtSignal, QSize, QRectF, QEvent, QPoint, QMimeData, QTimer
-from PyQt6.QtGui import QIcon, QColor, QPainter, QPen, \
-    QPainterPath, QPaintEvent, QAction, QResizeEvent, QEnterEvent, QDragEnterEvent
+from PyQt6.QtCore import Qt, pyqtSignal, QSize, QEvent, QPoint, QMimeData, QTimer
+from PyQt6.QtGui import QIcon, QColor, QAction, QResizeEvent, QEnterEvent, QDragEnterEvent
 from PyQt6.QtWidgets import QWidget, QToolButton, QPushButton, QTextEdit, QDialog, QApplication, QMessageBox
 from overrides import overrides
 from qtanim import fade_in
-from qthandy import pointy, gc, translucent, bold, clear_layout, decr_font, \
-    margins, spacer, sp, curved_flow, incr_icon, vbox, vspacer, transparent, underline
-from qthandy.filter import OpacityEventFilter, ObjectReferenceMimeData, DragEventFilter, DropEventFilter
+from qthandy import pointy, gc, translucent, decr_font, \
+    margins, spacer, sp, incr_icon, vbox, vspacer, transparent, underline
+from qthandy.filter import OpacityEventFilter, ObjectReferenceMimeData, DropEventFilter
 from qtmenu import ScrollableMenuWidget, ActionTooltipDisplayMode, MenuWidget, TabularGridMenuWidget
 
 from src.main.python.plotlyst.common import RELAXED_WHITE_COLOR
 from src.main.python.plotlyst.core.domain import Novel, Scene, SceneStructureItemType, SceneStructureItem, SceneOutcome, \
-    SceneStructureAgenda, ScenePurposeType
+    ScenePurposeType
 from src.main.python.plotlyst.env import app_env
 from src.main.python.plotlyst.view.common import action, fade_out_and_gc, ButtonPressResizeEventFilter
 from src.main.python.plotlyst.view.generated.scene_structure_editor_widget_ui import Ui_SceneStructureWidget
@@ -46,8 +45,8 @@ from src.main.python.plotlyst.view.icons import IconRegistry
 from src.main.python.plotlyst.view.style.base import apply_white_menu
 from src.main.python.plotlyst.view.widget.button import DotsMenuButton
 from src.main.python.plotlyst.view.widget.display import StageRecommendationBadge
-from src.main.python.plotlyst.view.widget.input import RemovalButton
 from src.main.python.plotlyst.view.widget.list import ListView, ListItemWidget
+from src.main.python.plotlyst.view.widget.outline import OutlineTimelineWidget, OutlineItemWidget
 from src.main.python.plotlyst.view.widget.scenes import SceneOutcomeSelector
 
 beat_descriptions = {SceneStructureItemType.BEAT: 'New action, reaction, thought, or emotion',
@@ -256,7 +255,8 @@ class BeatSelectorMenu(TabularGridMenuWidget):
         self._addAction(self._tabMisc, 'False victory', SceneStructureItemType.FALSE_VICTORY, 2, 0)
         self.addWidget(self._tabMisc, vspacer(), 6, 0)
 
-    def _addAction(self, tabWidget: QWidget, text: str, beat_type: SceneStructureItemType, row: int, column: int) -> QAction:
+    def _addAction(self, tabWidget: QWidget, text: str, beat_type: SceneStructureItemType, row: int,
+                   column: int) -> QAction:
         description = beat_descriptions[beat_type]
         action_ = action(text, beat_icon(beat_type), slot=lambda: self.selected.emit(beat_type), tooltip=description)
         self._actions[beat_type] = action_
@@ -307,43 +307,17 @@ class _PlaceholderWidget(QWidget):
         self.layout().addWidget(self.btn)
 
 
-class SceneStructureItemWidget(QWidget):
+class SceneStructureItemWidget(OutlineItemWidget):
     SceneBeatMimeType: str = 'application/scene-beat'
-    dragStarted = pyqtSignal()
-    dragStopped = pyqtSignal()
-    removed = pyqtSignal(object)
-    iconFixedSize: int = 36
 
     def __init__(self, novel: Novel, scene_structure_item: SceneStructureItem, parent=None, readOnly: bool = False):
-        super(SceneStructureItemWidget, self).__init__(parent)
+        super(SceneStructureItemWidget, self).__init__(scene_structure_item, parent, readOnly=readOnly)
         self.novel = novel
-        self._readOnly = readOnly
         self.beat = scene_structure_item
-        vbox(self, 0, 2)
 
-        self._btnName = QPushButton(self)
-        bold(self._btnName)
-        if app_env.is_mac():
-            self._btnName.setFixedHeight(max(self._btnName.sizeHint().height() - 8, 24))
-
-        self._btnIcon = QToolButton(self)
-        self._btnIcon.setIconSize(QSize(24, 24))
-        self._btnIcon.setFixedSize(self.iconFixedSize, self.iconFixedSize)
-
-        self._btnRemove = RemovalButton(self)
-        self._btnRemove.setHidden(True)
-        self._btnRemove.clicked.connect(self._remove)
-
-        self.layout().addWidget(self._btnIcon, alignment=Qt.AlignmentFlag.AlignCenter)
-        self.layout().addWidget(self._btnName, alignment=Qt.AlignmentFlag.AlignCenter)
-
-        if not self._readOnly:
-            self._btnIcon.setCursor(Qt.CursorShape.OpenHandCursor)
-            self._dragEventFilter = DragEventFilter(self, self.SceneBeatMimeType, self._beatDataFunc,
-                                                    grabbed=self._btnIcon, startedSlot=self.dragStarted.emit,
-                                                    finishedSlot=self.dragStopped.emit)
-            self._btnIcon.installEventFilter(self._dragEventFilter)
-            self.setAcceptDrops(True)
+    @overrides
+    def mimeType(self):
+        return self.SceneBeatMimeType
 
     def isEmotion(self) -> bool:
         return self.beat.type == SceneStructureItemType.EMOTION
@@ -358,24 +332,6 @@ class SceneStructureItemWidget(QWidget):
     @abstractmethod
     def copy(self) -> 'SceneStructureItemWidget':
         pass
-
-    @overrides
-    def resizeEvent(self, event: QResizeEvent) -> None:
-        self._btnRemove.setGeometry(self.width() - 15, self.iconFixedSize, 15, 15)
-        self._btnRemove.raise_()
-
-    @overrides
-    def enterEvent(self, event: QEnterEvent) -> None:
-        if not self._readOnly:
-            self._btnRemove.setVisible(True)
-
-    @overrides
-    def leaveEvent(self, event: QEvent) -> None:
-        self._btnRemove.setHidden(True)
-
-    def _remove(self):
-        anim = qtanim.fade_out(self, duration=150)
-        anim.finished.connect(lambda: self.removed.emit(self))
 
     def _initStyle(self):
         color = self._color()
@@ -395,8 +351,6 @@ class SceneStructureItemWidget(QWidget):
             padding-right: 15px;
         }}''')
 
-    def _beatDataFunc(self, btn):
-        return id(self)
 
     def _color(self) -> str:
         if self.beat.type == SceneStructureItemType.ACTION:
@@ -635,22 +589,13 @@ class SceneStructureEmotionWidget(SceneStructureItemWidget):
         self._btnIcon.setIcon(icon)
 
 
-class SceneStructureTimeline(QWidget):
+class SceneStructureTimeline(OutlineTimelineWidget):
     emotionChanged = pyqtSignal()
     outcomeChanged = pyqtSignal()
-    timelineChanged = pyqtSignal()
 
     def __init__(self, parent=None):
         super().__init__(parent)
-        self._novel: Optional[Novel] = None
         self._scene: Optional[Scene] = None
-        self._readOnly: bool = False
-        sp(self).h_exp().v_exp()
-        curved_flow(self, margin=10, spacing=10)
-
-        self._agenda: Optional[SceneStructureAgenda] = None
-        self._structure: List[SceneStructureItem] = []
-        self._beatWidgets: List[SceneStructureItemWidget] = []
 
         self._currentPlaceholder: Optional[QWidget] = None
         self._menuEmotions = EmotionSelectorMenu()
@@ -666,31 +611,13 @@ class SceneStructureTimeline(QWidget):
 
         self.setAcceptDrops(True)
 
-    def setNovel(self, novel: Novel):
-        self._novel = novel
-
     def setScene(self, scene: Scene):
         self._scene = scene
 
-    def setReadnOnly(self, readOnly: bool):
-        self._readOnly = readOnly
-
+    @overrides
     def clear(self):
-        self._beatWidgets.clear()
-        clear_layout(self)
+        super().clear()
         self._selectorMenu.setOutcomeEnabled(True)
-
-    def setStructure(self, items: List[SceneStructureItem]):
-        self.clear()
-
-        self._structure = items
-
-        for item in items:
-            self._addBeatWidget(item)
-        if not items:
-            self.layout().addWidget(self._newPlaceholderWidget(displayText=True))
-
-        self.update()
 
     def refreshOutcome(self):
         for wdg in self._beatWidgets:
@@ -698,83 +625,24 @@ class SceneStructureTimeline(QWidget):
                 if wdg.hasOutcome():
                     wdg.setOutcome(self._scene.outcome)
 
-    @overrides
-    def paintEvent(self, event: QPaintEvent) -> None:
-        painter = QPainter(self)
-        painter.setRenderHint(QPainter.RenderHint.Antialiasing)
-        painter.setOpacity(0.5)
-
-        pen = QPen()
-        pen.setColor(QColor('grey'))
-
-        pen.setWidth(3)
-        painter.setPen(pen)
-
-        path = QPainterPath()
-
-        forward = True
-        y = 0
-        for i, wdg in enumerate(self._beatWidgets):
-            pos: QPoint = wdg.pos()
-            pos.setY(pos.y() + wdg.layout().contentsMargins().top())
-            if isinstance(wdg, SceneStructureItemWidget):
-                pos.setY(pos.y() + wdg.iconFixedSize // 2)
-            pos.setX(pos.x() + wdg.layout().contentsMargins().left())
-            if i == 0:
-                y = pos.y()
-                path.moveTo(pos.toPointF())
-                painter.drawLine(pos.x(), y - 10, pos.x(), y + 10)
-            else:
-                if pos.y() > y:
-                    if forward:
-                        path.arcTo(QRectF(pos.x() + wdg.width(), y, 60, pos.y() - y),
-                                   90, -180)
-                    else:
-                        path.arcTo(QRectF(pos.x(), y, 60, pos.y() - y), -270, 180)
-                    forward = not forward
-                    y = pos.y()
-
-            if forward:
-                pos.setX(pos.x() + wdg.width())
-            path.lineTo(pos.toPointF())
-
-        painter.drawPath(path)
-        if self._beatWidgets:
-            if forward:
-                x_arrow_diff = -10
-            else:
-                x_arrow_diff = 10
-            painter.drawLine(pos.x(), y, pos.x() + x_arrow_diff, y + 10)
-            painter.drawLine(pos.x(), y, pos.x() + x_arrow_diff, y - 10)
-
     def _addBeat(self, beatType: SceneStructureItemType):
-        item = SceneStructureItem(beatType)
+        item = SceneStructureItem(type=beatType)
         if beatType == SceneStructureItemType.CLIMAX:
             item.outcome = SceneOutcome.DISASTER
         self._structure.append(item)
         self._addBeatWidget(item)
-
-    def _addBeatWidget(self, item: SceneStructureItem):
-        widget = self._newBeatWidget(item)
-        self._beatWidgets.append(widget)
-        if self.layout().count() == 0:
-            self.layout().addWidget(self._newPlaceholderWidget())
-        self.layout().addWidget(widget)
-        self.layout().addWidget(self._newPlaceholderWidget())
-        widget.activate()
-        self.timelineChanged.emit()
 
     def _insertBeat(self, beatType: SceneStructureItemType):
         if beatType == SceneStructureItemType.EMOTION:
             self._menuEmotions.exec(self.mapToGlobal(self._currentPlaceholder.pos()))
             return
 
-        item = SceneStructureItem(beatType)
+        item = SceneStructureItem(type=beatType)
         widget = self._newBeatWidget(item)
         self._insertWidget(item, widget)
 
     def _insertEmotion(self, emotion: str):
-        item = SceneStructureItem(SceneStructureItemType.EMOTION, emotion=emotion)
+        item = SceneStructureItem(type=SceneStructureItemType.EMOTION, emotion=emotion)
         widget = self._newBeatWidget(item)
         self._insertWidget(item, widget)
 
@@ -971,7 +839,7 @@ class SceneStructureList(ListView):
 
     @overrides
     def _addNewItem(self):
-        beat = SceneStructureItem(SceneStructureItemType.EXPOSITION)
+        beat = SceneStructureItem(type=SceneStructureItemType.EXPOSITION)
         self._items.append(beat)
         self.addItem(beat)
 
@@ -1143,9 +1011,9 @@ class SceneStructureTemplateSelector(QDialog, Ui_SceneStructuteTemplateSelector)
     def _fillInSceneTemplate(self):
         self._structure.clear()
         self._structure.extend([
-            SceneStructureItem(SceneStructureItemType.ACTION),
-            SceneStructureItem(SceneStructureItemType.CONFLICT),
-            SceneStructureItem(SceneStructureItemType.CLIMAX)
+            SceneStructureItem(type=SceneStructureItemType.ACTION),
+            SceneStructureItem(type=SceneStructureItemType.CONFLICT),
+            SceneStructureItem(type=SceneStructureItemType.CLIMAX)
         ])
 
         self.textBrowser.setText('Scene template')
@@ -1153,9 +1021,9 @@ class SceneStructureTemplateSelector(QDialog, Ui_SceneStructuteTemplateSelector)
     def _fillInSequelTemplate(self):
         self._structure.clear()
         self._structure.extend([
-            SceneStructureItem(SceneStructureItemType.REACTION),
-            SceneStructureItem(SceneStructureItemType.DILEMMA),
-            SceneStructureItem(SceneStructureItemType.DECISION)
+            SceneStructureItem(type=SceneStructureItemType.REACTION),
+            SceneStructureItem(type=SceneStructureItemType.DILEMMA),
+            SceneStructureItem(type=SceneStructureItemType.DECISION)
         ])
 
         self.textBrowser.setText('Sequel template')
